+++ pnpm-workspace.yaml
@@ -42,6 +42,7 @@ catalog:
   "luxon": ^3.5.0
   "lucide-react": ^0.475.0
   "vitest": ^1.5.1
+  "emittery": ^1.2.0
   # Hyperlane
   "viem": ^2.22.1
   "next-plausible": ^3.12.4
